extract command operations into @agent-crm/sdk (phase 3)
Migrate every CLI command file to a thin commander shell that calls a
pure SDK operation. The eight operations now live under
packages/sdk/src/operations/, each taking a Workspace and a structured
input and returning a structured result; never touching process.argv /
process.cwd / process.exit / process.stderr / process.stdin /
process.env.

* execute        -> query(workspace, sql, params) + dumpSchema(workspace)
* init           -> createWorkspace(absolutePath); seed helpers (which the
                    test fixture also uses) move alongside it
* schema         -> createObject / addAttribute / editAttributeOptions
* import-transcript -> importTranscript(workspace, payload) +
                    parseTranscriptPayload(raw)
* import-linkedin / -x / -post -> importLinkedinProfile /
                    importXProfile / importPost (each takes the Apify
                    token + cache dir as inputs; CLI shell reads env)
* import (CSV)   -> importCsv(workspace, { csvText, source, on*
                    callbacks for progress/flush events })
* records        -> planDedupe / applyDedupe / dedupeRecords /
                    createRecord / updateRecord. Dedup conflict
                    resolution is a static "keep"/"discard" or an async
                    ConflictResolver callback; CLI passes a readline
                    prompt for `--prefer interactive`.
* auth           -> already pure (PKCE primitives + token cache moved in
                    phase 1); the interactive flow stays in CLI per the
                    issue spec.

Add Workspace.fromLix() as an escape hatch so tests can hand an
existing Lix to the new SDK operations without rewriting their lix
plumbing.

The compat shim at packages/cli/src/workspace/open.ts is dead -
deleted. CLI workspace handling is now uniformly Workspace.open(
resolveWorkspacePath(opts.workspace)).

CLI surface (commands, flags, JSON output) is byte-identical; all 132
tests still pass; the SDK + CLI tarballs install and run together
end-to-end.

Refs #64.
