+++ packages/sdk/src/index.ts
@@ -30,3 +30,13 @@ export * from "./lib/uuidv7.js";
 
 export * from "./workspace/schemas/index.js";
 export { Workspace } from "./workspace.js";
+
+export * from "./operations/execute.js";
+export * from "./operations/init.js";
+export * from "./operations/import-csv.js";
+export * from "./operations/import-linkedin.js";
+export * from "./operations/import-post.js";
+export * from "./operations/import-transcript.js";
+export * from "./operations/import-x.js";
+export * from "./operations/records.js";
+export * from "./operations/schema.js";
